+++ package.js
@@ -1,6 +1,6 @@
 Package.describe({
   name: 'itgenio:live-modules',
-  version: '0.0.4',
+  version: '0.0.5',
   summary: 'Deploy any code to client\'s runtime',
   documentation: 'README.md',
   git: 'https://github.com/itgenio/meteor-live-modules.git',
@@ -27,5 +27,5 @@ Package.onTest(function(api) {
 
   api.addFiles(['tests/_init.ts', 'tests/utils.ts'], ['client', 'server']);
 
-  api.addFiles(['tests/evaluateAsCSS.ts', 'tests/client.ts'], 'client');
+  api.addFiles(['tests/evaluateAsCSS.ts', 'tests/liveModules.ts', 'tests/getModulesByNameOrTag.ts'], 'client');
 });