+++ src/getModulesByNameOrTag.ts
@@ -1,0 +1,5 @@
+import { LiveModulesCollection } from './shared';
+
+export function getModulesByNameOrTag(tagOrName: string) {
+  return LiveModulesCollection.find({ $or: [{ name: tagOrName }, { tags: tagOrName }] }).fetch();
+}