+++ tests/liveModules.ts
@@ -1,4 +1,4 @@
-import { LiveModules } from '../src/client';
+import { LiveModules } from '../src/liveModules';
 import { MODULE_NAME_REQUIRED, MODULE_NAME_NON_REQUIRED } from './_init';
 import { requireModule } from '../src/requireModule';
 
@@ -23,8 +23,21 @@ Tinytest.addAsync(
 Tinytest.addAsync(
   'LiveModules - importModule - by name',
   async function(test) {
-    await LiveModules.importModule(MODULE_NAME_NON_REQUIRED);
+    const m = await LiveModules.importModule(MODULE_NAME_NON_REQUIRED);
 
     test.isTrue(Object[MODULE_NAME_NON_REQUIRED]);
+    test.isUndefined(m);
+
+    const { name } = await LiveModules.importModule(MODULE_NAME_REQUIRED);
+    test.equal(name, MODULE_NAME_REQUIRED);
+  },
+);
+
+Tinytest.addAsync(
+  'LiveModules - require - by name',
+  async function(test) {
+    const m = LiveModules.require(MODULE_NAME_REQUIRED);
+    test.isNotUndefined(m);
+    test.equal(m.name, MODULE_NAME_REQUIRED);
   },
 );