+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+**0.0.5**
+
+- New: `LiveModules.require(tagOrName:string)`;
+- Changed: `LiveModules.importModule(): Promise<any>` now returns the module if it marks as required.
+
 **0.0.4**
 
 - New option `subOnStartup` to enable/disable subscribe on startup;